+++ wire-ios/Wire-iOS/Sources/UserInterface/StartUI/StartUI/StartUIViewController.swift
@@ -160,7 +160,6 @@ final class StartUIViewController: UIViewController {
         searchResults.searchResultsView.collectionView.accessibilityIdentifier = "search.list"
 
         createConstraints()
-        //updateActionBar()
         searchResults.searchContactList()
 
         view.accessibilityViewIsModal = true
